RHEL-Backport-Release: RHEL_86_5949_PRERELEASE_220707.3
Generation Information:
DII-TAG: DII_5949_hotfix3
Backport Branch: redhat/main
Backports Head: 1523337c

Hotfix: Update with changes as per DII 5.15 base.

DII has updated the base version to 5.15, which has resulted in
the relocation or deletion of a few files. This was not taken into
account during the update, resulting in compilation errors due to
duplicate definitions.

Signed-off-by: Nitin Gote <nitin.r.gote@intel.com>